Add support for a default page if not provided
If no page is specified in the request, e.g. the root of a website, then a default page can be configured to be
served up. This defaults to the index.html page within the bucket as per website standards if the
S3_HTTP_DEFAULT_PAGE environment variable has not been set.